+++ src/app/features/events/event-schedule-dialog/event-schedule-dialog.component.ts
@@ -1,4 +1,3 @@
-/* eslint-disable no-debugger */
 import { Component, Inject, OnInit, OnDestroy } from '@angular/core';
 import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
 import { FormBuilder, FormControl, FormGroup, Validators } from '@angular/forms';